requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
MetaTrader5==5.0.45
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
